Add Garlic Emote and Tin Can Emote
This commit also changes the sizing on emotes
from using 15px width to 15px height. Since both
the garlic and tin-can emotes utilize max height
on their canvas, this ensures consistency in size.

If an emote were to ever utilize max width on its
canvas, in that case we would use 15px width as
the measurement